+++ lms/lms/doctype/loan/loan.py
@@ -504,26 +504,17 @@ class Loan(Document):
         try:
             self.total_collateral_value = 0
             drawing_power = 0
-            if self.instrument_type == "Shares":
-                for i in self.items:
-                    i.amount = i.price * i.pledged_quantity
-                    self.total_collateral_value += i.amount
 
-                drawing_power = round(
-                    (self.total_collateral_value * (self.allowable_ltv / 100)), 2
-                )
-            else:  # for Drawing power Calculation
-                for i in self.items:
-                    i.amount = i.price * i.pledged_quantity
-                    i.eligible_amount = (i.eligible_percentage / 100) * i.amount
-                    self.total_collateral_value += i.amount
-                    drawing_power += i.eligible_amount
+            for i in self.items:
+                i.amount = i.price * i.pledged_quantity
+                i.eligible_amount = ((i.eligible_percentage or 0) / 100) * i.amount
+                self.total_collateral_value += i.amount
+                drawing_power += i.eligible_amount
 
-                drawing_power = round(
-                    drawing_power,
-                    2,
-                )
-
+            drawing_power = round(
+                drawing_power,
+                2,
+            )
             self.drawing_power = (
                 drawing_power
                 if drawing_power <= self.sanctioned_limit
@@ -547,9 +538,9 @@ class Loan(Document):
         # sauce: https://stackoverflow.com/a/23827026/9403680
         sql = """
 			SELECT
-				cl.loan, cl.isin, cl.psn, cl.pledgor_boid, cl.pledgee_boid , cl.prf, cl.scheme_code, cl.folio, cl.amc_code,
-				s.price, s.security_name,
-                (select sc.category_name from `tabSecurity Category` sc  where sc.name = als.security_category) as security_category
+				cl.loan, cl.isin, cl.psn, cl.pledgor_boid, cl.pledgee_boid, cl.prf, cl.scheme_code, cl.folio, cl.amc_code,
+				s.price, s.security_name,als.eligible_percentage,
+                als.category_name as security_category
 				, SUM(COALESCE(CASE WHEN request_type = 'Pledge' THEN quantity END,0))
 				- SUM(COALESCE(CASE WHEN request_type = 'Unpledge' THEN quantity END,0))
 				- SUM(COALESCE(CASE WHEN request_type = 'Sell Collateral' THEN quantity END,0)) quantity
@@ -621,7 +612,6 @@ class Loan(Document):
                 isin_folio_combo = "{}{}".format(i.isin, i.folio if i.folio else "")
                 curr = collateral_list_map.get(isin_folio_combo)
                 # curr = collateral_list_map.get(i.isin)
-                # print(check, i.price, curr.price, not check or i.price != curr.price)
                 if (not check or i.price != curr.price) and i.pledged_quantity > 0:
                     check = True
                     self.update_collateral_ledger(curr.price, curr.isin)
@@ -641,11 +631,12 @@ class Loan(Document):
                         "security_category": i.security_category,
                         "pledged_quantity": i.quantity,
                         "price": i.price,
+                        "eligible_percentage": i.eligible_percentage,
                     }
                 )
 
                 self.append("items", loan_item)
-
+            # update_ltv function merge with this
             return check
         except Exception:
             frappe.log_error(
@@ -663,7 +654,6 @@ class Loan(Document):
         # setting check flag
         for i in self.items:
             curr = collateral_list_map.get(i.isin)
-            # print(check, i.price, curr.price, not check or i.price != curr.price)
             if (not check or i.price != curr.price) and i.pledged_quantity > 0:
                 check = True
                 self.update_collateral_ledger(curr.price, curr.isin)
@@ -707,6 +697,8 @@ class Loan(Document):
             old_total_collateral_value = self.total_collateral_value
 
             check = self.update_items()
+            if on_approval == False:
+                self.update_ltv()
             if check:
                 self.fill_items()
                 self.available_topup_amt = self.max_topup_amount()
@@ -1786,6 +1778,37 @@ class Loan(Document):
         self.drawing_power_str = lms.amount_formatter(self.drawing_power)
         self.sanctioned_limit_str = lms.amount_formatter(self.sanctioned_limit)
         self.balance_str = lms.amount_formatter(self.balance)
+        for items in self.items:
+            if self.instrument_type == "Shares":
+                image = frappe.get_all(
+                    "Allowed Security",
+                    filters={
+                        "isin": items.isin,
+                    },
+                    fields=["amc_image"],
+                )
+                code = ""
+            else:
+                code = frappe.get_all(
+                    "Allowed Security",
+                    filters={
+                        "isin": items.isin,
+                    },
+                    fields=["amc_code"],
+                )
+                image = frappe.get_all(
+                    "AMC Master",
+                    filters={
+                        "name": code[0].amc_code,
+                    },
+                    fields=["amc_image"],
+                )
+
+            if image and image[0].amc_image:
+                amc_image = frappe.utils.get_url(image[0].amc_image)
+                items.amc_image = amc_image
+            if code:
+                items.amc_code = code[0].amc_code
 
     def save_loan_sanction_history(self, agreement_file, event="New loan"):
         loan_sanction_history = frappe.get_doc(
@@ -1820,23 +1843,29 @@ class Loan(Document):
                     actual_drawing_power += i.eligible_amount
             max_topup_amount = actual_drawing_power - self.sanctioned_limit
         else:
-            max_topup_amount = (
-                self.total_collateral_value * (self.allowable_ltv / 100)
-            ) - self.sanctioned_limit
+            # max_topup_amount = (
+            #     self.total_collateral_value * (self.allowable_ltv / 100)
+            # ) - self.sanctioned_limit
+            for i in self.items:
+                i.amount = i.price * i.pledged_quantity
+                i.eligible_amount = ((i.eligible_percentage or 0) / 100) * i.amount
+                total_collateral_value += i.amount
+                actual_drawing_power += i.eligible_amount
+
+            max_topup_amount = actual_drawing_power - self.sanctioned_limit
 
         # show available top up amount only if topup amount is greater than 10% of sanctioned limit
         # if self.instrument_type == "Mutual Fund":  # for Mutual Fund Topup
         if self.sanctioned_limit > lender.maximum_sanctioned_limit:
             max_topup_amount = 0
         elif (
-            (actual_drawing_power / self.sanctioned_limit * 100) - 100 >= 10
-            and max_topup_amount >= 1000
-            and self.instrument_type == "Mutual Fund"
-        ) or (
-            max_topup_amount > (self.sanctioned_limit * 0.1)
-            and max_topup_amount >= 1000
-            and self.instrument_type == "Shares"
-        ):
+            actual_drawing_power / self.sanctioned_limit * 100
+        ) - 100 >= 10 and max_topup_amount >= 1000:
+            # ) or (
+            #     actual_drawing_power > (self.sanctioned_limit * 0.1)
+            #     and max_topup_amount >= 1000
+            #     and self.instrument_type == "Shares"
+            # ):
             if (
                 max_topup_amount + self.sanctioned_limit
             ) > lender.maximum_sanctioned_limit:
@@ -1893,23 +1922,23 @@ class Loan(Document):
                 topup_doc.db_set("top_up_amount", 0)
             frappe.db.commit()
 
-    def max_unpledge_amount(self):
-        if self.instrument_type == "Shares":
-            minimum_collateral_value = (100 / self.allowable_ltv) * self.balance
-            maximum_unpledge_amount = (
-                self.total_collateral_value - minimum_collateral_value
-            )
+    # def max_unpledge_amount(self):
+    #     if self.instrument_type == "Shares":
+    #         minimum_collateral_value = (100 / self.allowable_ltv) * self.balance
+    #         maximum_unpledge_amount = (
+    #             self.total_collateral_value - minimum_collateral_value
+    #         )
 
-            return {
-                "minimum_collateral_value": round(minimum_collateral_value, 2)
-                if minimum_collateral_value > 0
-                else 0.0,
-                "maximum_unpledge_amount": round(maximum_unpledge_amount, 2)
-                if maximum_unpledge_amount > 0
-                else 0.0,
-            }
-        else:
-            return {}
+    #         return {
+    #             "minimum_collateral_value": round(minimum_collateral_value, 2)
+    #             if minimum_collateral_value > 0
+    #             else 0.0,
+    #             "maximum_unpledge_amount": round(maximum_unpledge_amount, 2)
+    #             if maximum_unpledge_amount > 0
+    #             else 0.0,
+    #         }
+    #     else:
+    #         return {}
 
     def update_pending_sell_collateral_amount(self):
         all_pending_sell_collateral_applications = frappe.get_all(
@@ -2225,6 +2254,15 @@ class Loan(Document):
             ),
             as_dict=1,
         )[0]["amount"]
+
+    def update_ltv(self):
+        for i in self.items:
+            i.eligible_percentage = frappe.db.get_value(
+                "Allowed Security",
+                {"isin": i.isin, "lender": self.lender},
+                "eligible_percentage",
+            )
+        return
 
 
 def check_loans_for_shortfall(loans):